refactor(ui): split Settings tab panels into children, drop grandfather (#855)
Extract the three lighter, independently-loaded tab panels — workspace, device
and diagnose — into self-loading children under settings/, leaving the heavy
session panel (30+ state vars, steersEl scroll) untouched in the parent shell.
Each child owns its panel's exclusive state, functions and onMount data fetch;
the parent keeps the panel wrappers (tab/hidden/aria) + the session panel.
Trade-off: getSettings() is now read twice on open (parent for session fields,
workspace child for currentRoot) — an idempotent GET, no UX change.

Settings <template> 59/60 -> clears the Tier-1 40/60 bar (0 findings; >=19
cyclomatic reduction); the three children are each well under it. The Settings
.fallowrc.jsonc grandfather is deleted.
